Reorganize integration tests based on Rust best practices.
Following the recommendations from the rust book:
https://doc.rust-lang.org/book/ch11-03-test-organization.html

This is in order to make integration tests simpler.

We can potentially make the things we share more strict, but it doesn't
really matter as this shouldn't be used as a lib.
