+++ server/svix-server/src/v1/endpoints/attempt.rs
@@ -40,18 +40,18 @@ use crate::v1::utils::Pagination;
 
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize, ModelOut)]
 #[serde(rename_all = "camelCase")]
-struct MessageAttemptOut {
-    response: String,
-    response_status_code: i16,
-    status: MessageStatus,
-    trigger_type: MessageAttemptTriggerType,
-    msg_id: MessageId,
-    endpoint_id: EndpointId,
+pub struct MessageAttemptOut {
+    pub response: String,
+    pub response_status_code: i16,
+    pub status: MessageStatus,
+    pub trigger_type: MessageAttemptTriggerType,
+    pub msg_id: MessageId,
+    pub endpoint_id: EndpointId,
 
-    id: MessageAttemptId,
+    pub id: MessageAttemptId,
 
     #[serde(rename = "timestamp")]
-    created_at: DateTime<Utc>,
+    pub created_at: DateTime<Utc>,
 }
 
 // FIXME: This can and should be a derive macro
@@ -75,12 +75,12 @@ impl From<messageattempt::Model> for MessageAttemptOut {
 /// that message.
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
-struct AttemptedMessageOut {
+pub struct AttemptedMessageOut {
     #[serde(flatten)]
-    msg: MessageOut,
-    status: MessageStatus,
+    pub msg: MessageOut,
+    pub status: MessageStatus,
     #[serde(skip_serializing_if = "Option::is_none")]
-    next_attempt: Option<DateTimeWithTimeZone>,
+    pub next_attempt: Option<DateTimeWithTimeZone>,
 }
 
 impl ModelOut for AttemptedMessageOut {
@@ -409,156 +409,4 @@ pub fn router() -> Router {
             .route("attempt/endpoint/:endp_id/", get(list_attempts_by_endpoint))
             .route("attempt/msg/:msg_id/", get(api_not_implemented)),
     )
-}
-
-#[cfg(test)]
-mod tests {
-    use reqwest::StatusCode;
-
-    use super::{AttemptedMessageOut, MessageAttemptOut};
-    use crate::{
-        test_util::{run_with_retries, start_svix_server, TestReceiver},
-        v1::{
-            endpoints::{
-                application::tests::create_test_app, endpoint::tests::create_test_endpoint,
-                message::tests::create_test_message,
-            },
-            utils::ListResponse,
-        },
-    };
-
-    #[tokio::test]
-    #[cfg_attr(not(feature = "integration_testing"), ignore)]
-    async fn test_list_attempted_messages() {
-        let (client, _jh) = start_svix_server();
-
-        let app_id = create_test_app(&client, "v1AttemptListAttemptedMessagesTestApp")
-            .await
-            .unwrap();
-
-        let receiver_1 = TestReceiver::start(axum::http::StatusCode::OK);
-        let receiver_2 = TestReceiver::start(axum::http::StatusCode::OK);
-
-        let endp_id_1 = create_test_endpoint(&client, &app_id, &receiver_1.endpoint)
-            .await
-            .unwrap();
-        let endp_id_2 = create_test_endpoint(&client, &app_id, &receiver_2.endpoint)
-            .await
-            .unwrap();
-
-        let msg_1 = create_test_message(&client, &app_id, serde_json::json!({"test": "data1"}))
-            .await
-            .unwrap();
-        let msg_2 = create_test_message(&client, &app_id, serde_json::json!({"test": "data2"}))
-            .await
-            .unwrap();
-        let msg_3 = create_test_message(&client, &app_id, serde_json::json!({"test": "data3"}))
-            .await
-            .unwrap();
-
-        let list_1: ListResponse<AttemptedMessageOut> = client
-            .get(
-                &format!("api/v1/app/{}/endpoint/{}/msg/", app_id, endp_id_1),
-                StatusCode::OK,
-            )
-            .await
-            .unwrap();
-        let list_2: ListResponse<AttemptedMessageOut> = client
-            .get(
-                &format!("api/v1/app/{}/endpoint/{}/msg/", app_id, endp_id_2),
-                StatusCode::OK,
-            )
-            .await
-            .unwrap();
-
-        for list in [list_1, list_2] {
-            assert_eq!(list.data.len(), 3);
-
-            // Assert order
-            assert_eq!(list.data[0].msg.id, msg_3);
-            assert_eq!(list.data[1].msg.id, msg_2);
-            assert_eq!(list.data[2].msg.id, msg_1);
-
-            let message_ids: Vec<_> = list.data.into_iter().map(|amo| amo.msg.id).collect();
-            assert!(message_ids.contains(&msg_1));
-            assert!(message_ids.contains(&msg_2));
-            assert!(message_ids.contains(&msg_3));
-        }
-    }
-
-    #[tokio::test]
-    #[cfg_attr(not(feature = "integration_testing"), ignore)]
-    async fn test_list_attempts_by_endpoint() {
-        let (client, _jh) = start_svix_server();
-
-        let app_id = create_test_app(&client, "v1AttemptListAttemptsByEndpointTestApp")
-            .await
-            .unwrap();
-
-        let receiver_1 = TestReceiver::start(axum::http::StatusCode::OK);
-        let receiver_2 = TestReceiver::start(axum::http::StatusCode::OK);
-
-        let endp_id_1 = create_test_endpoint(&client, &app_id, &receiver_1.endpoint)
-            .await
-            .unwrap();
-        let endp_id_2 = create_test_endpoint(&client, &app_id, &receiver_2.endpoint)
-            .await
-            .unwrap();
-
-        let msg_1 = create_test_message(&client, &app_id, serde_json::json!({"test": "data1"}))
-            .await
-            .unwrap();
-        let msg_2 = create_test_message(&client, &app_id, serde_json::json!({"test": "data2"}))
-            .await
-            .unwrap();
-        let msg_3 = create_test_message(&client, &app_id, serde_json::json!({"test": "data3"}))
-            .await
-            .unwrap();
-
-        // And wait at most one second for all attempts to be processed
-        run_with_retries(|| async {
-            for endp_id in [endp_id_1.clone(), endp_id_2.clone()] {
-                let list: ListResponse<MessageAttemptOut> = client
-                    .get(
-                        &format!("api/v1/app/{}/attempt/endpoint/{}/", app_id, endp_id),
-                        StatusCode::OK,
-                    )
-                    .await
-                    .unwrap();
-
-                if list.data.len() != 3 {
-                    anyhow::bail!("list len {}, not 3", list.data.len());
-                }
-            }
-
-            Ok(())
-        })
-        .await
-        .unwrap();
-
-        let list_1: ListResponse<MessageAttemptOut> = client
-            .get(
-                &format!("api/v1/app/{}/attempt/endpoint/{}/", app_id, endp_id_1),
-                StatusCode::OK,
-            )
-            .await
-            .unwrap();
-        let list_2: ListResponse<MessageAttemptOut> = client
-            .get(
-                &format!("api/v1/app/{}/attempt/endpoint/{}/", app_id, endp_id_2),
-                StatusCode::OK,
-            )
-            .await
-            .unwrap();
-
-        for list in [list_1, list_2] {
-            let message_ids: Vec<_> = list.data.into_iter().map(|amo| amo.msg_id).collect();
-            assert!(message_ids.contains(&msg_1));
-            assert!(message_ids.contains(&msg_2));
-            assert!(message_ids.contains(&msg_3));
-        }
-
-        receiver_1.jh.abort();
-        receiver_2.jh.abort();
-    }
 }
