+++ server/svix-server/src/v1/endpoints/endpoint/secrets.rs
@@ -95,80 +95,3 @@ pub(super) async fn rotate_endpoint_secret(
 
     Ok((StatusCode::NO_CONTENT, Json(EmptyResponse {})))
 }
-
-#[cfg(test)]
-pub mod tests {
-    use reqwest::StatusCode;
-
-    use crate::{
-        test_util::{start_svix_server, IgnoredResponse},
-        v1::endpoints::{
-            application::tests::create_test_app,
-            endpoint::{tests::post_endpoint_default, EndpointSecretOut},
-        },
-    };
-
-    // Simply tests that upon rotating an endpoint secret that it differs from the prior one
-    #[tokio::test]
-    #[cfg_attr(not(feature = "integration_testing"), ignore)]
-    async fn test_endpoint_secret_get_and_rotation() {
-        let (client, _jh) = start_svix_server();
-
-        const APP_NAME: &str = "v1EndpointSecretRotationTestApp";
-        const EP_URI: &str = "http://v1EndpointSecretRotationTestEp.test";
-
-        let app_id = create_test_app(&client, APP_NAME).await.unwrap();
-
-        let ep = post_endpoint_default(&client, &app_id, EP_URI)
-            .await
-            .unwrap();
-
-        let former_secret: EndpointSecretOut = client
-            .get(
-                &format!("api/v1/app/{}/endpoint/{}/secret/", app_id, ep.id),
-                StatusCode::OK,
-            )
-            .await
-            .unwrap();
-
-        let _: IgnoredResponse = client
-            .post(
-                &format!("api/v1/app/{}/endpoint/{}/secret/rotate/", app_id, ep.id),
-                serde_json::json!({ "key": null }),
-                StatusCode::NO_CONTENT,
-            )
-            .await
-            .unwrap();
-
-        assert_ne!(
-            former_secret,
-            client
-                .get(
-                    &format!("api/v1/app/{}/endpoint/{}/secret/", app_id, ep.id),
-                    StatusCode::OK
-                )
-                .await
-                .unwrap()
-        );
-
-        let _: IgnoredResponse = client
-            .post(
-                &format!("api/v1/app/{}/endpoint/{}/secret/rotate/", app_id, ep.id),
-                &former_secret,
-                StatusCode::NO_CONTENT,
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(
-            former_secret,
-            client
-                .get(
-                    &format!("api/v1/app/{}/endpoint/{}/secret/", app_id, ep.id),
-                    StatusCode::OK
-                )
-                .await
-                .unwrap()
-        );
-    }
-}
